backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
pydantic>=2.9.0
pydantic[email]>=2.9.0
python-dotenv>=1.0.0
httpx>=0.25.2
google-generativeai>=0.3.2
pgvector>=0.2.4

